Move content from Core project to OpenTelemetry.AutoInstrumentation
