+++ OpenTelemetry.AutoInstrumentation.sln
@@ -56,8 +56,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Examples.ConsoleApp.SelfBoo
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Examples.OpenTracingLibrary", "examples\OpenTracingLibrary\Examples.OpenTracingLibrary.csproj", "{7220BDBB-084D-40B3-A710-19E3C7147185}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenTelemetry.AutoInstrumentation.Core", "src\OpenTelemetry.AutoInstrumentation.Core\OpenTelemetry.AutoInstrumentation.Core.csproj", "{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Examples.BindingRedirect", "examples\BindingRedirect\Examples.BindingRedirect.csproj", "{94089CFF-FD2A-45EF-A496-874569882E81}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Examples.OldReference", "examples\OldReference\Examples.OldReference.csproj", "{4C10E5C1-E212-4744-B92E-49A3AFC314FC}"
@@ -243,18 +241,6 @@ Global
 		{7220BDBB-084D-40B3-A710-19E3C7147185}.Release|x64.Build.0 = Release|Any CPU
 		{7220BDBB-084D-40B3-A710-19E3C7147185}.Release|x86.ActiveCfg = Release|Any CPU
 		{7220BDBB-084D-40B3-A710-19E3C7147185}.Release|x86.Build.0 = Release|Any CPU
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}.Debug|x64.Build.0 = Debug|Any CPU
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}.Debug|x86.Build.0 = Debug|Any CPU
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}.Release|x64.ActiveCfg = Release|Any CPU
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}.Release|x64.Build.0 = Release|Any CPU
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}.Release|x86.ActiveCfg = Release|Any CPU
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100}.Release|x86.Build.0 = Release|Any CPU
 		{94089CFF-FD2A-45EF-A496-874569882E81}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{94089CFF-FD2A-45EF-A496-874569882E81}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{94089CFF-FD2A-45EF-A496-874569882E81}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -598,7 +584,6 @@ Global
 		{6690361C-CCD8-42C4-A5CF-82D554B03A3D} = {9E5F0022-0A50-40BF-AC6A-C3078585ECAB}
 		{7EA51D34-8E0E-4548-8A7F-5C6F9C3533AF} = {BAF8F246-3645-42AD-B1D0-0F7EAFBAB34A}
 		{7220BDBB-084D-40B3-A710-19E3C7147185} = {BAF8F246-3645-42AD-B1D0-0F7EAFBAB34A}
-		{8BA1CCCC-B07C-46DD-9D54-0B98B3001100} = {9E5F0022-0A50-40BF-AC6A-C3078585ECAB}
 		{94089CFF-FD2A-45EF-A496-874569882E81} = {BAF8F246-3645-42AD-B1D0-0F7EAFBAB34A}
 		{4C10E5C1-E212-4744-B92E-49A3AFC314FC} = {BAF8F246-3645-42AD-B1D0-0F7EAFBAB34A}
 		{5728056A-51AA-4FF5-AD0C-E86E44E36102} = {5C915382-C886-457D-8641-9E766D8E5A17}
