--- src/OpenTelemetry.AutoInstrumentation.Core/OpenTelemetry.AutoInstrumentation.Core.csproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <RootNamespace>OpenTelemetry.AutoInstrumentation</RootNamespace>
-    <AssemblyName>OpenTelemetry.AutoInstrumentation.Core</AssemblyName>
-
-    <!-- NuGet -->
-    <Version>0.3.0</Version>
-    <IsPackable>false</IsPackable>
-
-    <!-- Allow the GenerateAssemblyInfo task in the .NET SDK to generate all
-         assembly attributes except for the AssemblyVersionAttribute. This will
-         be locked at Major.0.0.0 and is defined in AssemblyInfo.cs -->
-    <GenerateAssemblyVersionAttribute>false</GenerateAssemblyVersionAttribute>
-  </PropertyGroup>
-
-</Project>
