+++ src/OpenTelemetry.AutoInstrumentation/ClrNames.cs
